refactor: useChatSocket 분리
